type modification fps

+++ inc/xeve.h
@@ -340,7 +340,7 @@ typedef struct _XEVE_PARAM
     /* height of input frame */
     int            h;
     /* frame rate (Hz) */
-    int            fps;
+    double         fps;
     /* MAX I-frame period in frames.
     - 0: only one I-frame at the first time.
     - 1: every frame is coded in I-frame
